+++ exonum-java-binding/core/src/main/java/com/exonum/binding/core/runtime/TransactionInvoker.java
@@ -19,11 +19,9 @@ package com.exonum.binding.core.runtime;
 import static com.google.common.base.Preconditions.checkArgument;
 
 import com.exonum.binding.core.service.Service;
+import com.exonum.binding.core.transaction.ExecutionException;
 import com.exonum.binding.core.transaction.TransactionContext;
-import com.exonum.binding.core.transaction.TransactionExecutionException;
 import com.google.inject.Inject;
-import java.lang.invoke.MethodHandle;
-import java.lang.invoke.WrongMethodTypeException;
 import java.util.Map;
 
 /**
@@ -49,30 +47,15 @@ final class TransactionInvoker {
    *
    * @throws IllegalArgumentException if there is no transaction method with given id in a
    *     corresponding service
-   * @throws TransactionExecutionException if {@link TransactionExecutionException} was thrown by
+   * @throws ExecutionException if {@link ExecutionException} was thrown by
    *     the transaction method, it is propagated
-   * @throws UnexpectedTransactionExecutionException if any other exception is thrown by
+   * @throws UnexpectedExecutionException if any other exception is thrown by
    *     the transaction method, it is wrapped as cause
    */
-  void invokeTransaction(int transactionId, byte[] arguments, TransactionContext context)
-      throws TransactionExecutionException {
+  void invokeTransaction(int transactionId, byte[] arguments, TransactionContext context) {
     checkArgument(transactionMethods.containsKey(transactionId),
         "No method with transaction id (%s)", transactionId);
     TransactionMethod transactionMethod = transactionMethods.get(transactionId);
-    Object argumentsObject = transactionMethod.serializeArguments(arguments);
-    MethodHandle methodHandle = transactionMethod.getMethodHandle();
-    try {
-      methodHandle.invoke(service, argumentsObject, context);
-    } catch (WrongMethodTypeException | ClassCastException invocationException) {
-      // Invocation-specific exceptions are thrown as is — they are not thrown
-      // from the _transaction method_, but from framework code (see mh#invoke spec).
-      throw invocationException;
-    } catch (TransactionExecutionException serviceException) {
-      // 'Service-defined' transaction exceptions
-      throw serviceException;
-    } catch (Throwable unexpectedServiceException) {
-      // Any other _transaction_ exceptions
-      throw new UnexpectedTransactionExecutionException(unexpectedServiceException);
-    }
+    transactionMethod.invoke(service, arguments, context);
   }
 }